fix: carousel component is now reusable
