+++ src/components/carousel/carousel.styles.js
@@ -1,5 +1,4 @@
 import styled from "styled-components";
-import { fontSize, lineHeight, space, width } from "styled-system";
 
 export const CarrouselCards = styled.div`
   display: grid;
@@ -12,7 +11,7 @@ export const CarrouselCards = styled.div`
   margin: 60px auto 30px auto;
   cursor: pointer;
   position: relative;
-  padding-bottom: 38px;
+  
 
   scroll-snap-type: both mandatory;
 
@@ -39,63 +38,13 @@ export const CarrouselCards = styled.div`
   }
 
   @media only screen and (max-width: 500px) {
-    width: 95vw;
+    width: 100%;
     margin: 40px 0 30px 0;
   }
-`;
 
-export const CarrouselCard = styled.div`
-  background: ${({ theme }) => theme.colors.gray};
-  width: 520px;
-  height: 240px;
-  border-radius: 20px;
-  overflow: hidden;
-  display: inline-block;
-  position: relative;
-  scroll-snap-align: center;
-  ${lineHeight};
-  ${space};
-  ${fontSize};
-  ${width};
-
-  @media only screen and (max-width: 1024px) {
-    display: flex;
-    flex-direction: column;
-    width: 434px;
-    height: 268px;
+  @media only screen and (max-width: 375px) {
+    padding: 0 10px 38px 10px;
   }
-
-  @media only screen and (max-width: 500px) {
-    width: 355px;
-    height: 288px;
-  }
-`;
-
-export const TitleReviewWrapper = styled.div`
-  display: flex;
-  justify-content: space-between;
-  ${lineHeight};
-  ${space};
-  ${fontSize};
-  ${width};
-`;
-
-export const ReviewText = styled.p`
-  font-family: ${({ theme }) => theme.fonts.roboto};
-
-  text-align: left;
-  ${lineHeight};
-  ${space};
-  ${fontSize};
-  ${width};
-`;
-
-export const Stars = styled.div`
-  display: flex;
-  justify-content: space-between;
-  position: absolute;
-  top: 20px;
-  right: 30px;
 `;
 
 export const AngleContainer = styled.div`
@@ -119,11 +68,23 @@ export const Angle = styled.a`
 
   &:first-child {
     left: 50px;
+    @media only screen and (max-width: 500px) {
+        left: 10px;
+  }
+    @media only screen and (max-width: 375px) {
+        left: 20px;
+}
   }
 
   &:nth-child(2) {
     align-self: flex-end;
     right: 50px;
     top: -17px;
+    @media only screen and (max-width: 500px) {
+        right: 10px;
+  }
+    @media only screen and (max-width: 375px) {
+        right: 20px;
+  }
   }
 `;
