+++ src/components/products/products.js
@@ -3,11 +3,8 @@ import React from 'react';
 import Image from "next/image";
 import { GlobalSection, GlobalSubtitle, GlobalTitle, SpanTitle } from "../global-components";
 import {
-  Angle,
   CarrouselCard,
-  CarrouselCards,
   ProductText,
-  AngleContainer,
   TitleWrapper,
   ImageWrapper,
   InfoWrapper,
@@ -15,51 +12,11 @@ import {
   Price
 } from "./products.styles";
 
+import Carousel from "../carousel/carousel";
+
 const Products = ({ data }) => {
-  const carouselRef = React.useRef();
-  const [percent, setPercent] = React.useState(0)
-
-
-  React.useEffect(() => {
-    carouselRef.current.addEventListener("scroll", e => {
-      setPercent(e.target.scrollLeft / (e.target.scrollWidth - e.target.offsetWidth))
-    })
-
-  }, [])
-
-  React.useEffect(() => {
-    if (carouselRef) {
-      let isDown = false;
-      let startX;
-      let scrollLeft;
-
-      carouselRef.current.addEventListener('mousedown', (e) => {
-        isDown = true;
-        startX = e.pageX - carouselRef.current.offsetLeft;
-        scrollLeft = carouselRef.current.scrollLeft;
-      })
-
-      carouselRef.current.addEventListener('mouseleave', () => {
-        isDown = false;
-      });
-
-      carouselRef.current.addEventListener('mouseup', () => {
-        isDown = false;
-      });
-
-      carouselRef.current.addEventListener('mousemove', (e) => {
-        if (!isDown) return;
-        e.preventDefault();
-        const x = e.pageX - carouselRef.current.offsetLeft;
-        const scrollX = (x - startX) * 1;
-
-        carouselRef.current.scrollLeft = scrollLeft - scrollX;
-      });
-    }
-
-  }, [])
-
   return (
+
     <GlobalSection px={[0, 40, 40, 0]} pt={[20, 60, 80]}>
       <GlobalTitle
         fontSize={[34, 40, 54, 50]}
@@ -70,7 +27,7 @@ const Products = ({ data }) => {
         <SpanTitle>beauty products</SpanTitle> for u {" "}
       </GlobalTitle>
 
-      <CarrouselCards ref={carouselRef}>
+      <Carousel>
         {data.map((products, index) => {
           return (
             <CarrouselCard mb={[60, 40]} key={index}>
@@ -105,31 +62,8 @@ const Products = ({ data }) => {
         })}
 
 
-      </CarrouselCards>
-
-      <AngleContainer>
-        <Angle
-          classname="prev"
-          onClick={e => carouselRef.current.scrollBy({
-            left: -100,
-            behavior: 'smooth'
-          })}
-          disabled={percent === 0}
-        >
-          <Image src="/LeftAngle.svg" width="20" height="20" intrinsic="true" />
-        </Angle>
-        <Angle
-          className="next"
-          onClick={e => carouselRef.current.scrollBy({
-            left: 100,
-            behavior: 'smooth'
-          })}
-          disabled={percent >= 1}
-        >
-          <Image src="/RightAngle.svg" width="20" height="20" intrinsic="true" />
-        </Angle>
-      </AngleContainer>
-
+      
+      </Carousel>
     </GlobalSection>
   );
 };
